ngambil seluruh fitur (#22)
* Revisi 05/04/2025

* Menambahkan Nama Kelompok beserta NIM Anggota Kelompok

* SCRUM-7 fitur dark theme

* SCRUM-7

* SCRUM-15 Commit Laravel dan Tampilan awal Dashboard dengan memberikan akses cepat ke fitur fitur utama

* SCRUM-13 #PBI 16

* SCRUM-13 PBI 16

* SCRUM-13 perbaikan routing

* SCRUM-14 Mengintegrasikan seluruh fitur SPRINT-1 ke dashboard

* SCRUM-14 PBI#17

* Revisi 04/05/2025

* resolve

* Merge Fitur Simulasi Hemat Energi ke Branch Utama (#12)

* SCRUM-9 Selesai Membuat Tampilan Grafik

* adding class information

* SCRUM-22 Membuat Tampilan Awal Page Fitur Rekomendasi Produk yang Menampilkan Daftar Produk Energi Terbarukan

* SCRUM-22 Membuat Tampilan Awal Page Fitur Rekomendasi Produk yang Menampilkan Daftar Produk Energi Terbarukan

* SCRUM-23 Melanjutkan PBI#1 hingga selesai dan Memulai Progres PBI#2 Menampilkan Ulasan Pengguna Pada Page Fitur Rekomendasi Produk

* SCRUM-24 Menyelesaikan PBI#2 dan Membuat Dasaran untuk PBI#3

* SCRUM-24 Menyelesaikan PBI#3 dengan membuat sistem rekomendasi produk pada bagian detail produk dan halaman utama fitur produk

* SCRUM-24 Menghapus controller yang tidak dipakai

* SCRUM-39 SCRUM-40 SCRUM-41 Membuat codingan awal untuk PBI #23, PBI #24, dan PBI #25 untuk Fitur Simulasi Hemat Energi

* SCRUM-39 Menghubungkan fitur dengan branch utama dan memperbaiki bug yang ada, sekaligus melakukan commit ulang untuk PBI #23 karena tidak terecord di jira

* SCRUM-40 Melakukan commit ulang untuk PBI #24 dan membuat folder backup

* SCRUM-41 Memperbaiki code secara keseluruhan terutama di PBI#25 Fitur Simulasi Hemat Energi: Memberikan prediksi penghematan biaya

---------


* Merge Sarah_Branch to main_branch  (#14)

* Penambahan keterangan kelas dan kelompok

* SCRUM-10

* PBI#7

* Delete enerzero directory

* SCRUM-16 PBI#7 Leaderboard

* SCRUM-18 PBI#12 Sistem Leaderboard

* SCRUM-17 PBI#11 Sistem Leaderboard

* SCRUM-16 PBI#10 Perbaikan tampilan

* SCRUM-48 PBI#29 Fitur Notifikasi

* SCRUM-48 perbaikan

* SCRUM-50 PBI#31 Notifikasi Artikel Baru

---------


* Merge fitur challenges ke main_branch (#15)

* Memasukan nim dalam data nama anggota PPL (proyek perangkat lunak)

* SCRUM-8

* Create README.md

* Update README.md

* SCRUM-19 PBI#4 Forum menampilkan semua thread atau percakapan tiap pengguna secara review

* SCRUM-19 PBI#4 Forum menampilkan semua thread atau percakapan tiap pengguna secara review

* SCRUM-19 PBI #7 memperbarui tampilan, menambahkan sedikit interaksi terhadap UI, memperbaiki sidebar menjadi fixed position

* SCRUM-21 SCRUM-20 PBI#8#9 commit perubahan pada UI, penambahan interaksi terhadap UI, menambahkan detail dari forum untuk PBI#8 dan menambahkan CRUD untuk PBI#9 pada forum, user yang melakukan post mempunyai akses terahadap forum yang ia buat (update, create, delete, read)

* SCRUM-20 SCRUM-21 menyambungkan data pada forum agar dapat ditampilkan ke dashboard, forum yang ditampilkan berdasarkan jumlah likenya, menambahkan tombol back pada detail forum

* Add authentication for edit, delete, and update in forum feature, only creator of the forum can do edit, delete, and update

* SCRUM-33 #PBI 26 pull dari main, dan push folder struktur untuk fitur mission/challenges

* SCRUM-33 membuat semua tampilan dari fitur tantangan ini, sehingga user dapat melihat bagaiaman tampilan awal dari fitur tantangan

* SCRUM-33 SCRUM-34 SCRUM-35 #PBI 26 Melihat daftar tantangan ramah lingkungan yang dapat dipilih; PBI #27 Menampilkan grafik progress dari misi; PBI #28 Sistem Menampilkan notifikasi keberhasilan misi

---------


* Menggabungkan fitur yang diperbaiki  yaitu progress bar pada onprogress fetaure di challenges (#17)

* Memasukan nim dalam data nama anggota PPL (proyek perangkat lunak)

* SCRUM-8

* Create README.md

* Update README.md

* SCRUM-19 PBI#4 Forum menampilkan semua thread atau percakapan tiap pengguna secara review

* SCRUM-19 PBI#4 Forum menampilkan semua thread atau percakapan tiap pengguna secara review

* SCRUM-19 PBI #7 memperbarui tampilan, menambahkan sedikit interaksi terhadap UI, memperbaiki sidebar menjadi fixed position

* SCRUM-21 SCRUM-20 PBI#8#9 commit perubahan pada UI, penambahan interaksi terhadap UI, menambahkan detail dari forum untuk PBI#8 dan menambahkan CRUD untuk PBI#9 pada forum, user yang melakukan post mempunyai akses terahadap forum yang ia buat (update, create, delete, read)

* SCRUM-20 SCRUM-21 menyambungkan data pada forum agar dapat ditampilkan ke dashboard, forum yang ditampilkan berdasarkan jumlah likenya, menambahkan tombol back pada detail forum

* Add authentication for edit, delete, and update in forum feature, only creator of the forum can do edit, delete, and update

* SCRUM-33 #PBI 26 pull dari main, dan push folder struktur untuk fitur mission/challenges

* SCRUM-33 membuat semua tampilan dari fitur tantangan ini, sehingga user dapat melihat bagaiaman tampilan awal dari fitur tantangan

* SCRUM-33 SCRUM-34 SCRUM-35 #PBI 26 Melihat daftar tantangan ramah lingkungan yang dapat dipilih; PBI #27 Menampilkan grafik progress dari misi; PBI #28 Sistem Menampilkan notifikasi keberhasilan misi

* SCRUM-34 memasukan fungsi progress bar pada onprogress challenges yang diambil, sisa menampilkan notifikasi

* membenarkan fitur progress bar pada onprogress challenges

---------


* menggabungkan fitur education dan update layout map (#19)

* Memasukan nim dalam data nama anggota PPL (proyek perangkat lunak)

* SCRUM-8

* Create README.md

* Update README.md

* SCRUM-19 PBI#4 Forum menampilkan semua thread atau percakapan tiap pengguna secara review

* SCRUM-19 PBI#4 Forum menampilkan semua thread atau percakapan tiap pengguna secara review

* SCRUM-19 PBI #7 memperbarui tampilan, menambahkan sedikit interaksi terhadap UI, memperbaiki sidebar menjadi fixed position

* SCRUM-21 SCRUM-20 PBI#8#9 commit perubahan pada UI, penambahan interaksi terhadap UI, menambahkan detail dari forum untuk PBI#8 dan menambahkan CRUD untuk PBI#9 pada forum, user yang melakukan post mempunyai akses terahadap forum yang ia buat (update, create, delete, read)

* SCRUM-20 SCRUM-21 menyambungkan data pada forum agar dapat ditampilkan ke dashboard, forum yang ditampilkan berdasarkan jumlah likenya, menambahkan tombol back pada detail forum

* Add authentication for edit, delete, and update in forum feature, only creator of the forum can do edit, delete, and update

* SCRUM-33 #PBI 26 pull dari main, dan push folder struktur untuk fitur mission/challenges

* SCRUM-33 membuat semua tampilan dari fitur tantangan ini, sehingga user dapat melihat bagaiaman tampilan awal dari fitur tantangan

* SCRUM-33 SCRUM-34 SCRUM-35 #PBI 26 Melihat daftar tantangan ramah lingkungan yang dapat dipilih; PBI #27 Menampilkan grafik progress dari misi; PBI #28 Sistem Menampilkan notifikasi keberhasilan misi

* SCRUM-34 memasukan fungsi progress bar pada onprogress challenges yang diambil, sisa menampilkan notifikasi

* membenarkan fitur progress bar pada onprogress challenges

* SCRUM-35 SCRUM-44 untuk scrum44 finalisasi dari fitur edukasi, untuk scrum35 finalisasi untuk notifikasi dari challenges

* memperbaiki layout pada map dan education

---------


* merge fitur notifikasi dengan fitur lainnya

* Penambahan keterangan kelas dan kelompok

* tes komit

* SCRUM-10

* SCRUM-11

* PBI#7

* Delete enerzero directory

* SCRUM-16 PBI#7 Leaderboard

* SCRUM-29 mebuat mcr fitur laporan dan analisis

* SCRUM-18 PBI#12 Sistem Leaderboard

* SCRUM-17 PBI#11 Sistem Leaderboard

* uji coba

* SCRUM-29 Membuat fitur laporan penggunaan energi

* SCRUM-16 PBI#10 Perbaikan tampilan

* SCRUM-29

* SCRUM #45 menampilkan profile

* SCRUM-48 PBI#29 Fitur Notifikasi

* SCRUM-48 perbaikan

* SCRUM 45 update

* SCRUM-50 PBI#31 Notifikasi Artikel Baru

* save local changes before pulling

* SCRUM-49 PBI#30 Menyambungkan fitur notifikasi dan fitur challenge

---------


---------

+++ enerzero/app/Http/Controllers/ForumController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Forum;
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\ForumReply;
 use App\Models\ForumLike;
@@ -10,6 +11,13 @@ use Illuminate\Support\Facades\Auth;
 
 class ForumController extends Controller
 {
+    
+    public function __construct()
+    {
+        $this->middleware('auth')->except(['index', 'show']);
+    }
+
+    
     // Menampilkan semua forum
     public function index(Request $request)
     {
@@ -54,12 +62,24 @@ class ForumController extends Controller
     public function edit($id)
     {
         $forum = Forum::findOrFail($id);
+        // cek user access
+        if ($forum->username !== Auth::user()->username) {
+            return back()->with('error', 'Kamu tidak memiliki izin untuk mengakses aksi ini.');
+        }
+        
         return view('forum.edit', compact('forum'));
     }
 
     // Update data forum
     public function update(Request $request, $id)
     {
+        $forum = Forum::findOrFail($id);
+        // cek user access
+        if ($forum->username !== Auth::user()->username) {
+            return back()->with('error', 'Kamu tidak memiliki izin untuk mengakses aksi ini.');
+        }
+        
+
         $request->validate([
             'title' => 'required|string|max:255',
             'description' => 'required|string',
@@ -75,6 +95,11 @@ class ForumController extends Controller
     public function destroy($id)
     {
         $forum = Forum::findOrFail($id);
+        // cek user access
+        if ($forum->username !== Auth::user()->username) {
+            return back()->with('error', 'Kamu tidak memiliki izin untuk mengakses aksi ini.');
+        }
+        
         $forum->delete();
 
         return redirect()->route('forum', $forum->id)->with('success', 'Forum berhasil dihapus!');
